+++ src/main/scala/jobs/users/AveragePurchasesUser.scala
@@ -74,7 +74,7 @@ class AveragePurchasesUser(sc: SparkContext) extends Actor with ActorLogging  wi
       classOf[BSONObject]
     ).filter((t: Tuple2[Object, BSONObject]) => {
       def parseFloat(d: String): Option[Long] = {
-        try { Some(d.toLong) } catch { case _: Throwable => None }
+        try { Some(d.toDouble.toLong) } catch { case _: Throwable => None }
       }
       parseFloat(t._2.get("lowerDate").toString) match {
         case Some(dbDate) => {
