+++ src/main/scala/jobs/users/PayingUsers.scala
@@ -111,7 +111,7 @@ class PayingUsers(
        classOf[BSONObject]
      ).filter((t: Tuple2[Object, BSONObject]) => {
        def parseFloat(d: String): Option[Long] = {
-         try { Some(d.toLong) } catch { case _: Throwable => None }
+         try { Some(d.toDouble.toLong) } catch { case _: Throwable => None }
        }
        parseFloat(t._2.get("time").toString) match {
          case Some(dbDate) => {
